requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
beautifulsoup4==4.9.1
Brotli==1.0.7
category-encoders==2.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.0
cycler==0.10.0
dash==1.14.0
dash-bootstrap-components==0.10.5
dash-core-components==1.10.2
dash-html-components==1.0.3
dash-renderer==1.6.0
dash-table==4.9.0
decorator==4.4.2
dill==0.3.2
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
lime==0.2.0.1
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.0.2
more-itertools==8.4.0
networkx==2.5
numpy==1.16.3
opencv-python==4.4.0.42
packaging==20.4
pandas==0.24.2
patsy==0.5.1
pdfcrowd==4.12.0
percy==2.0.2
pickle-mixin==1.0.2
Pillow==7.2.0
plotly==4.9.0
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.0.1
pytest-mock==3.2.0
pytest-sugar==0.9.4
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
requests==2.24.0
retrying==1.3.3
scikit-image==0.17.2
scikit-learn==0.22.2.post1
scipy==1.5.2
seaborn==0.9.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
statsmodels==0.11.1
termcolor==1.1.0
tifffile==2020.8.13
toml==0.10.1
tqdm==4.48.2
urllib3==1.25.10
waitress==1.4.4
Werkzeug==1.0.1
xgboost==0.81
zipp==3.1.0
